--- packages/app/src/components/presentational/FyghterRenamingModal/index.tsx
@@ -1,1 +1,0 @@
-export * from "./FyghterRenamingModal";
