+++ packages/app/src/components/presentational/RenameFyghterButton/RenameFyghterForm.tsx
@@ -1,0 +1,52 @@
+import React from "react";
+import { Modal, Form, Input } from "antd";
+
+// interface Values {}
+
+interface Props {
+  visible: boolean;
+  // TODO: To use Values type above
+  // See more: https://github.com/ant-design/ant-design/issues/21195
+  onSave: (values: {}) => void;
+  onCancel: () => void;
+  name: string;
+}
+
+export const RenameFyghterForm: React.FC<Props> = ({ visible, onCancel, onSave, name }) => {
+  const [form] = Form.useForm();
+  return (
+    <Modal
+      visible={visible}
+      title="Rename fyghter"
+      okText="Save"
+      onCancel={onCancel}
+      onOk={async (): Promise<void> => {
+        const values = await form.validateFields();
+        form.resetFields();
+        onSave(values);
+      }}
+    >
+      <Form
+        form={form}
+        layout="vertical"
+        name="form_in_modal"
+        initialValues={{
+          name,
+        }}
+      >
+        <Form.Item
+          name="name"
+          label="Name"
+          rules={[
+            {
+              required: true,
+              message: "Please input the name of the fyghter!",
+            },
+          ]}
+        >
+          <Input />
+        </Form.Item>
+      </Form>
+    </Modal>
+  );
+};
