+++ packages/app/src/components/presentational/RenameFyghterButton/index.tsx
@@ -1,0 +1,1 @@
+export * from "./RenameFyghterButton";
